+++ src/components/dropdown/dropdown.ts
@@ -255,7 +255,7 @@ export default class SlDropdown extends LitElement {
     // Other keys bring focus to the menu and initiate type-to-select behavior
     const ignoredKeys = ['Tab', 'Shift', 'Meta', 'Ctrl', 'Alt'];
     if (this.open && !ignoredKeys.includes(event.key)) {
-      menu.typeToSelect(event.key);
+      menu.typeToSelect(event);
     }
   }